Remove merge strategies baloney
