+++ .rubocop_todo.yml
@@ -176,7 +176,6 @@ RSpec/ContextWording:
     - 'spec/active_document/criteria/queryable/extensions/integer_spec.rb'
     - 'spec/active_document/criteria/queryable/extensions/time_spec.rb'
     - 'spec/active_document/criteria/queryable/extensions/time_with_zone_spec.rb'
-    - 'spec/active_document/criteria/queryable/mergeable_spec.rb'
     - 'spec/active_document/criteria/queryable/optional_spec.rb'
     - 'spec/active_document/criteria/queryable/selectable_logical_spec.rb'
     - 'spec/active_document/criteria/queryable/selectable_where_spec.rb'
