+++ lib/active_document/criteria/queryable/extensions/hash.rb
@@ -8,90 +8,6 @@ module ActiveDocument
         # Adds query type-casting behavior to Hash class.
         module Hash
 
-          # Add an object to a hash using the merge strategies.
-          #
-          # @example Add an object to a hash.
-          #   { field: value }.__add__({ field: other_value })
-          #
-          # @param [ Hash ] object The other hash to add.
-          #
-          # @return [ Hash ] The hash with object added.
-          def __add__(object)
-            apply_strategy(:__add__, object)
-          end
-
-          # Merge this hash into the provided array.
-          #
-          # @example Merge the hash into the array.
-          #   { field: value }.__add_from_array__([ 1, 2 ])
-          #
-          # @param [ Array ] array The array to add to.
-          #
-          # @return [ Hash ] The merged hash.
-          def __add_from_array__(array)
-            { keys.first => array.__add__(values.first) }
-          end
-
-          # Add an object to a hash using the merge strategies.
-          #
-          # @example Add an object to a hash.
-          #   { field: value }.__intersect__({ field: other_value })
-          #
-          # @param [ Hash ] object The other hash to intersect.
-          #
-          # @return [ Hash ] The hash with object intersected.
-          def __intersect__(object)
-            apply_strategy(:__intersect__, object)
-          end
-
-          # Merge this hash into the provided array.
-          #
-          # @example Merge the hash into the array.
-          #   { field: value }.__intersect_from_array__([ 1, 2 ])
-          #
-          # @param [ Array ] array The array to intersect to.
-          #
-          # @return [ Hash ] The merged hash.
-          def __intersect_from_array__(array)
-            { keys.first => array.__intersect__(values.first) }
-          end
-
-          # Merge this hash into the provided object.
-          #
-          # @example Merge the hash into the object.
-          #   { field: value }.__intersect_from_object__([ 1, 2 ])
-          #
-          # @param [ Object ] object The object to intersect to.
-          #
-          # @return [ Hash ] The merged hash.
-          def __intersect_from_object__(object)
-            { keys.first => object.__intersect__(values.first) }
-          end
-
-          # Add an object to a hash using the merge strategies.
-          #
-          # @example Add an object to a hash.
-          #   { field: value }.__union__({ field: other_value })
-          #
-          # @param [ Hash ] object The other hash to union.
-          #
-          # @return [ Hash ] The hash with object unioned.
-          def __union__(object)
-            apply_strategy(:__union__, object)
-          end
-
-          # Merge this hash into the provided object.
-          #
-          # @example Merge the hash into the object.
-          #   { field: value }.__union_from_object__([ 1, 2 ])
-          #
-          # @param [ Object ] object The object to union to.
-          #
-          # @return [ Hash ] The merged hash.
-          def __union_from_object__(object)
-            { keys.first => object.__union__(values.first) }
-          end
-
           # Make a deep copy of this hash.
           #
           # @example Make a deep copy of the hash.
@@ -116,27 +32,6 @@ module ActiveDocument
             tap do |hash|
               hash.each_pair do |key, value|
                 hash.store(key, ActiveDocument::Criteria::Translator.to_direction(value))
-              end
-            end
-          end
-
-          private
-
-          # Apply the provided strategy for the hash with the given object.
-          #
-          # @api private
-          #
-          # @example Apply the strategy.
-          #   { field: value }.apply_strategy(:__add__, 1)
-          #
-          # @param [ Symbol ] strategy The strategy to apply.
-          # @param [ Object ] object The object to merge.
-          #
-          # @return [ Hash ] The merged hash.
-          def apply_strategy(strategy, object)
-            tap do |hash|
-              object.each_pair do |key, value|
-                hash.store(key, hash[key].send(strategy, value))
               end
             end
           end
