+++ lib/active_document/criteria/queryable/extensions/object.rb
@@ -8,90 +8,6 @@ module ActiveDocument
         # Adds query type-casting behavior to Object class.
         module Object
 
-          # Combine the two objects using the add strategy.
-          #
-          # @example Add the object to the array.
-          #   [ 1, 2, 3 ].__add__(4)
-          #
-          # @param [ Object ] object The object to add.
-          #
-          # @return [ Object ] The result of the add.
-          def __add__(object)
-            object == self ? self : [self, object].flatten.uniq
-          end
-
-          # Merge this object into the provided array.
-          #
-          # @example Merge the object into the array.
-          #   4.__add_from_array__([ 1, 2 ])
-          #
-          # @param [ Array ] array The array to add to.
-          #
-          # @return [ Array ] The merged object.
-          def __add_from_array__(array)
-            array.concat(Array(self)).uniq
-          end
-
-          # Combine the two objects using the intersect strategy.
-          #
-          # @example Add the object to the array.
-          #   [ 1, 2, 3 ].__intersect__(4)
-          #
-          # @param [ Object ] object The object to intersect.
-          #
-          # @return [ Array ] The result of the intersect.
-          def __intersect__(object)
-            object.__intersect_from_object__(self)
-          end
-
-          # Merge this object into the provided array.
-          #
-          # @example Merge the object into the array.
-          #   4.__intersect_from_array__([ 1, 2 ])
-          #
-          # @param [ Array ] array The array to intersect to.
-          #
-          # @return [ Array ] The merged object.
-          def __intersect_from_array__(array)
-            array & Array(self)
-          end
-
-          # Merge this object into the provided array.
-          #
-          # @example Merge the object into the array.
-          #   4.__intersect_from_object__([ 1, 2 ])
-          #
-          # @param [ Object ] object The value to intersect to.
-          #
-          # @return [ Array ] The merged object.
-          def __intersect_from_object__(object)
-            Array(object) & Array(self)
-          end
-
-          # Combine the two objects using the union strategy.
-          #
-          # @example Add the object to the array.
-          #   [ 1, 2, 3 ].__union__(4)
-          #
-          # @param [ Object ] object The object to union.
-          #
-          # @return [ Array ] The result of the union.
-          def __union__(object)
-            object.__union_from_object__(self)
-          end
-
-          # Merge this object into the provided array.
-          #
-          # @example Merge the object into the array.
-          #   4.__union_from_object__([ 1, 2 ])
-          #
-          # @param [ Object ] object The value to union to.
-          #
-          # @return [ Array ] The merged object.
-          def __union_from_object__(object)
-            (Array(object) + Array(self)).uniq
-          end
-
           # Deep copy the object. This is for API compatibility, but needs to be
           # overridden.
           #
