+++ spec/active_document/criteria/queryable/extensions/array_spec.rb
@@ -3,6 +3,7 @@
 require 'spec_helper'
 
 describe Array do
+
   describe '.evolve' do
 
     context 'when provided an Array' do
@@ -42,65 +43,7 @@ describe Array do
     end
   end
 
-  describe '#__add__' do
-
-    context 'when the other object is a non-enumerable' do
-
-      context 'when all values are unique' do
-
-        it 'returns an array of both' do
-          expect([5, 6].__add__(7)).to eq([5, 6, 7])
-        end
-      end
-
-      context 'when values are not unique' do
-
-        it 'returns a unique array of both' do
-          expect([5, 6].__add__(6)).to eq([5, 6])
-        end
-      end
-    end
-
-    context 'when the other object is an array' do
-
-      context 'when all values are unique' do
-
-        it 'returns an array of both' do
-          expect([4, 5].__add__([6, 7])).to eq([4, 5, 6, 7])
-        end
-      end
-
-      context 'when values are not unique' do
-
-        it 'returns a unique array of both' do
-          expect([4, 5].__add__([5, 6, 7])).to eq([4, 5, 6, 7])
-        end
-      end
-    end
-
-    context 'when the other object is a hash' do
-
-      context 'when all values are unique' do
-
-        it 'returns a hash of both' do
-          expect([4, 5].__add__({ '$in' => [6, 7] })).to eq(
-            { '$in' => [4, 5, 6, 7] }
-          )
-        end
-      end
-
-      context 'when values are not unique' do
-
-        it 'returns a unique array of both' do
-          expect([4, 5].__add__({ '$in' => [5, 6, 7] })).to eq(
-            { '$in' => [4, 5, 6, 7] }
-          )
-        end
-      end
-    end
-  end
-
-  describe '__deep_copy__' do
+  describe '#__deep_copy__' do
 
     let(:inner) do
       [3, 4]
@@ -210,7 +153,7 @@ describe Array do
     end
   end
 
-  describe '__evolve_time__' do
+  describe '#__evolve_time__' do
 
     context 'when the array is composed of times' do
 
@@ -301,64 +244,6 @@ describe Array do
     end
   end
 
-  describe '#__intersect__' do
-
-    context 'when the other object is a non-enumerable' do
-
-      context 'when the values intersect' do
-
-        it 'returns an intersected array' do
-          expect([5, 6, 7].__intersect__(6)).to eq([6])
-        end
-      end
-
-      context 'when the values do not intersect' do
-
-        it 'returns an empty array' do
-          expect([5, 6, 7].__intersect__(8)).to be_empty
-        end
-      end
-    end
-
-    context 'when the other object is an array' do
-
-      context 'when the values intersect' do
-
-        it 'returns an intersected array' do
-          expect([4, 5, 6].__intersect__([6, 7])).to eq([6])
-        end
-      end
-
-      context 'when values do not intersect' do
-
-        it 'returns an empty array' do
-          expect([4, 5].__intersect__([6, 7])).to be_empty
-        end
-      end
-    end
-
-    context 'when the other object is a hash' do
-
-      context 'when the values intersect' do
-
-        it 'returns a hash of both' do
-          expect([4, 5].__intersect__({ '$in' => [5, 6] })).to eq(
-            { '$in' => [5] }
-          )
-        end
-      end
-
-      context 'when values do not intersect' do
-
-        it 'returns an empty array' do
-          expect([4, 5].__intersect__({ '$in' => [6, 7] })).to eq(
-            { '$in' => [] }
-          )
-        end
-      end
-    end
-  end
-
   describe '#__sort_option__' do
 
     context 'when the array is one dimensional' do
@@ -389,32 +274,6 @@ describe Array do
 
       it 'returns the array as a hash' do
         expect([:field, 1].__sort_pair__).to eq({ field: 1 })
-      end
-    end
-  end
-
-  describe '#__union__' do
-
-    context 'when the other object is a non-enumerable' do
-
-      it 'returns an unioned array' do
-        expect([5, 6, 7].__union__(6)).to eq([5, 6, 7])
-      end
-    end
-
-    context 'when the other object is an array' do
-
-      it 'returns an unioned array' do
-        expect([4, 5, 6].__union__([6, 7])).to eq([4, 5, 6, 7])
-      end
-    end
-
-    context 'when the other object is a hash' do
-
-      it 'returns a hash with a unioned array' do
-        expect([4, 5].__union__({ '$in' => [5, 6] })).to eq(
-          { '$in' => [4, 5, 6] }
-        )
       end
     end
   end
